+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add Terraform support via WASM plugin
 - Add Zig support via WASM plugin
 
 - [0.16.0]